+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - awsx.ecs.Cluster can be created from an existing aws.ecs.Cluster's id.
 - Add OPTIONS as a valid method and add ability to set custom gateway responses for
   [awsx.apigateway.API].
+- Load balancing targets can now be simply added to an ALB, NLB, Listener or TargetGroup using the
+  new `.attachTarget` methods on the respective classes.
 
 ### Compatibility issues